Es modules with "define" syntax instead of classes (#527)
* "sequelize.define" syntax with es modules

Applying the same models to multiple databases at the same time causes problems when the class syntax is used that comes along with the "-l esm" option. So I added another option "-l esmd" which creates esm6 modules, but with the sequelize.define syntax instead of the class inhereitance syntax.
Background: Since the init methods are static, applying the models to a second/third/... sequelize instance with the class syntax will change the sequelize-instance in all the previously initialized models. (Even when dynamic imports are used to "reload" the model files; javascript is too smart and does not load them again.)  This does not happen with the define syntax, since  there the models are only applied to the involved sequelize-instance.

* extend the types

Previously I forgot to extend the allowed loangOptions (I did my best to implement my changes working in javascript in typescript, but at least forgot to extend the allowed LangOptions.)